separate the traj and armature visualization to separate file
